fix(createEvent): refactor method (#168)
Redefine definitions for modifyEventWithPrompt.
