+++ src/index.ts
@@ -11,9 +11,11 @@ import type { CalendarSource } from "./schemas/interfaces/calendar-source";
 import { CalendarSourceType } from "./schemas/enums/calendar-source-type";
 import { CalendarType } from "./schemas/enums/calendar-type";
 import type { CapacitorCalendarPlugin } from "./definitions";
-import type { CreateEventWithPromptOptions } from "./sub-definitions/event-operations";
+import type { CreateEventWithPromptOptions } from "./schemas/interfaces/create-event-with-prompt-options";
 import { EventAvailability } from "./schemas/enums/event-availability";
+import type { EventEditAction } from "./schemas/types/event-edit-action";
 import { EventSpan } from "./schemas/enums/event-span";
+import type { ModifyEventWithPromptOptions } from "./schemas/interfaces/modify-event-with-prompt-options";
 import type { Reminder } from "./schemas/interfaces/reminder";
 import { ReminderRecurrenceFrequency } from "./schemas/enums/reminder-recurrence-frequency";
 import type { ReminderRecurrenceRule } from "./schemas/interfaces/reminder-recurrence-rule";
@@ -36,6 +38,8 @@ export type {
   CheckAllPermissionsResult,
   RequestAllPermissionsResult,
   CreateEventWithPromptOptions,
+  ModifyEventWithPromptOptions,
+  EventEditAction,
 };
 export {
   CalendarChooserSelectionStyle,
