+++ src/schemas/interfaces/modify-event-with-prompt-options.ts
@@ -1,0 +1,14 @@
+import { CreateEventWithPromptOptions } from "./create-event-with-prompt-options";
+
+/**
+ * @since 7.1.0
+ */
+export interface ModifyEventWithPromptOptions
+  extends CreateEventWithPromptOptions {
+  /**
+   * The ID of the event to be modified.
+   *
+   * @since 7.1.0
+   */
+  id: string;
+}
