+++ src/schemas/types/event-edit-action.ts
@@ -1,0 +1,4 @@
+/**
+ * @since 7.1.0
+ */
+export type EventEditAction = "canceled" | "saved" | "deleted";
